python/requirements.txt
flask
redis
urllib3
beautifulsoup4
selenium
wikipedia
mysql-connector-python
